manageiq_tags: refactor ManageIQTags class out to utils
